Move Login to containers
